refactor: move toggle expand all action to dedicated route file
